BUGFIX: gain tags written by mp3gain were never read correctly. Normalization using mp3gain was therefore impossible. Didn't anyone ever use that feature?!
Fixed a German translation.

git-svn-id: https://gtkpod.svn.sf.net/svnroot/gtkpod/gtkpod/trunk@413 f01d2545-417e-4e96-918e-98f8d0dbbcb6

+++ src/misc.h
@@ -1,4 +1,4 @@
-/* Time-stamp: <2004-07-19 22:25:47 jcs>
+/* Time-stamp: <2004-07-22 00:32:03 jcs>
 |
 |  Copyright (C) 2002-2003 Jorg Schuler <jcsjcs at users.sourceforge.net>
 |  Part of the gtkpod project.
@@ -136,5 +136,5 @@ void do_selected_tracks (void (*do_func)(GList *trackids));
 void do_selected_entry (void (*do_func)(GList *trackids), gint inst);
 void do_selected_playlist (void (*do_func)(GList *trackids));
 
-gint32 replaygain_to_soundcheck (gint gain);
+gint32 replaygain_to_soundcheck (gdouble gain);
 #endif
